+++ CHANGELOG.md
@@ -27,6 +27,7 @@ all of them.
  - Add extension `VK_KHR_external_fence_win32` (revision 1)
  - Add extension `VK_KHR_external_fence_fd` (revision 1)
  - Add extension `VK_KHR_variable_pointers` (revision 1)
+ - Add extension `VK_KHR_dedicated_allocation` (revision 1)
 
 ### Breaking changes